+++ tests/integration_tests/test_metric_internal.py
