+++ tests/integration_tests/test_start_method.py
