+++ tests/integration_tests/test_system_info.py
