+++ tests/integration_tests/test_telemetry_full.py
