+++ tests/unit_tests/test_model_workflows.py
@@ -29,10 +29,3 @@ def test_add_any():
             _add_any(artifact, ["invalid input type"], "invalid")
 
     assert True
-
-
-def test_offline_link_artifact(wandb_init):
-    run = wandb_init(mode="offline")
-    with pytest.raises(NotImplementedError):
-        run.link_artifact(None, "entity/project/portfolio", "latest")
-    run.finish()
